+++ sdk/src/org.graalvm.nativeimage/src/org/graalvm/nativeimage/impl/HeapDumpSupport.java
@@ -44,7 +44,4 @@ public interface HeapDumpSupport {
 
     void dumpHeap(String outputFile, boolean live) throws java.io.IOException;
 
-    void dumpHeapOnOutOfMemoryError();
-
-    void initHeapDumpOnOutOfMemoryErrorPath();
 }