+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/os/RawFileOperationSupport.java
@@ -29,6 +29,7 @@ import java.io.File;
 import org.graalvm.compiler.api.replacements.Fold;
 import org.graalvm.nativeimage.ImageSingletons;
 import org.graalvm.nativeimage.c.type.CCharPointer;
+import org.graalvm.nativeimage.impl.UnmanagedMemorySupport;
 import org.graalvm.word.Pointer;
 import org.graalvm.word.UnsignedWord;
 import org.graalvm.word.WordBase;
@@ -72,6 +73,15 @@ public interface RawFileOperationSupport {
     }
 
     /**
+     * Tries to allocate a platform-dependent C string for the given path. Note that the returned
+     * value needs to be freed manually once it is no longer needed (see
+     * {@link UnmanagedMemorySupport#free}).
+     *
+     * @return If the allocation is successful, a non-null value is returned.
+     */
+    CCharPointer allocateCPath(String path);
+
+    /**
      * Creates a file with the specified {@link FileCreationMode creation} and {@link FileAccessMode
      * access modes}.
      *
@@ -96,6 +106,7 @@ public interface RawFileOperationSupport {
      * @return If the operation is successful, it returns the file descriptor. Otherwise, it returns
      *         a value where {@link #isValid} will return false.
      */
+    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     RawFileDescriptor create(CCharPointer path, FileCreationMode creationMode, FileAccessMode accessMode);
 
     /**
@@ -120,6 +131,7 @@ public interface RawFileOperationSupport {
      * @return If the operation is successful, it returns the file descriptor. Otherwise, it returns
      *         a value where {@link #isValid} will return false.
      */
+    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     RawFileDescriptor open(CCharPointer path, FileAccessMode accessMode);
 
     /**